+++ packages/jest-cli/src/reporters/notify_reporter.js
@@ -11,6 +11,7 @@ import type {GlobalConfig} from 'types/Config';
 import type {AggregatedResult} from 'types/TestResult';
 import type {Context} from 'types/Context';
 
+import exit from 'exit';
 import path from 'path';
 import util from 'util';
 import notifier from 'node-notifier';
@@ -78,12 +79,41 @@ export default class NotifyReporter extends BaseReporter {
         result.numTotalTests,
       );
 
-      notifier.notify({
-        icon,
-        message,
-        title,
-      });
+      const watchMode = this._globalConfig.watch || this._globalConfig.watchAll;
+      const restartAnswer = 'Run again';
+      const quitAnswer = 'Exit tests';
+
+      if (!watchMode) {
+        notifier.notify({
+          icon,
+          message,
+          title,
+        });
+      } else {
+        notifier.notify(
+          {
+            actions: [restartAnswer, quitAnswer],
+            closeLabel: 'Close',
+            icon,
+            message,
+            title,
+          },
+          (err, _, metadata) => {
+            if (err || !metadata) {
+              return;
+            }
+            if (metadata.activationValue === quitAnswer) {
+              exit(0);
+              return;
+            }
+            if (metadata.activationValue === restartAnswer) {
+              this._startRun(this._globalConfig);
+            }
+          },
+        );
+      }
     }
+
     this._context.previousSuccess = success;
     this._context.firstRun = false;
   }