+++ CHANGELOG.md
@@ -37,8 +37,8 @@
 
 ### Fixes
 
-* `[jest-cli]` Remove the notifier actions in case of failure.
-  ([#5861](https://github.com/facebook/jest/pull/5861))
+* `[jest-cli]` Remove the notifier actions in case of failure when not in watch
+  mode. ([#5861](https://github.com/facebook/jest/pull/5861))
 * `[jest-runner]` Assign `process.env.JEST_WORKER_ID="1"` when in runInBand mode
   ([#5860](https://github.com/facebook/jest/pull/5860))
 * `[jest-cli]` Add descriptive error message when trying to use